net: ethernet: lantiq_etop: take interrupt from devicetree
The current version of this driver uses hardcoded interrupt numbers.
Interrupt allocation does not work since kernel 6.1 and returns error
code -89. This patch fixes this problem by allocating interrupts based
on devicetree.

The driver allocates as many queues as the number of interrupts
defined in dts.

The maximum number of queues depends on the SoC geeration:
- amazon-se -> 4 RX, 2 TX,
- danube -> 4 RX, 4 TX.

Signed-off-by: Aleksander Jan Bajkowski <olek2@wp.pl>